+++ core/dom/document.js
@@ -43,6 +43,8 @@ CKEDITOR.tools.extend( CKEDITOR.dom.document.prototype, {
 	 * @param {String} cssFileUrl The CSS file URL.
 	 */
 	appendStyleSheet: function( cssFileUrl ) {
+		cssFileUrl = CKEDITOR.appendTimestamp( cssFileUrl );
+
 		if ( this.$.createStyleSheet )
 			this.$.createStyleSheet( cssFileUrl );
 		else {